more fixture refactor
