+++ shared-data/js/__tests__/fixtures.test.ts
@@ -24,8 +24,6 @@ import {
   getCutoutConfigReplacmentForModule,
   getFixtureDisplayName,
   getFlexStackerD3Compatibility,
-  getMainNonComboFixtureId,
-  getMainUsbModuleFixtureIdForComboFixture,
   getVisualSlotIdForAA,
   getWasteChuteComboFixture,
   isFixtureInUsbModules,
@@ -373,83 +371,6 @@ describe('getFixtureDisplayName', () => {
       STAGING_AREA_SLOT_WITH_MAGNETIC_BLOCK_V1_FIXTURE
     )
     expect(result).toEqual('deck_configuration:module_with_staging_area')
-  })
-})
-
-describe('getMainModuleFixtureIdForComboFixture', () => {
-  it('should get flex stacker module fixture id', () => {
-    const result = getMainUsbModuleFixtureIdForComboFixture([
-      FLEX_STACKER_WITH_MAG_BLOCK_FIXTURE,
-      FLEX_STACKER_V1_FIXTURE,
-      FLEX_STACKER_WITH_WASTE_CHUTE_ADAPTER_NO_COVER_FIXTURE,
-    ])
-    expect(result).toEqual('flexStackerModuleV1')
-  })
-
-  it('should return null when no usb module is found', () => {
-    const result = getMainUsbModuleFixtureIdForComboFixture([
-      WASTE_CHUTE_RIGHT_ADAPTER_NO_COVER_FIXTURE,
-      WASTE_CHUTE_RIGHT_ADAPTER_COVERED_FIXTURE,
-    ])
-    expect(result).toEqual(null)
-  })
-})
-
-describe('getMainNonComboFixtureId', () => {
-  it('should get main non combo fixture id for waste chute', () => {
-    const result = getMainNonComboFixtureId(
-      [
-        WASTE_CHUTE_RIGHT_ADAPTER_COVERED_FIXTURE,
-        WASTE_CHUTE_RIGHT_ADAPTER_NO_COVER_FIXTURE,
-        STAGING_AREA_SLOT_WITH_WASTE_CHUTE_RIGHT_ADAPTER_COVERED_FIXTURE,
-        STAGING_AREA_SLOT_WITH_WASTE_CHUTE_RIGHT_ADAPTER_NO_COVER_FIXTURE,
-        FLEX_STACKER_WITH_WASTE_CHUTE_ADAPTER_NO_COVER_FIXTURE,
-      ],
-      ['gripperWasteChute'],
-      'cutoutD3'
-    )
-    expect(result).toEqual(WASTE_CHUTE_RIGHT_ADAPTER_NO_COVER_FIXTURE)
-  })
-
-  it('should get main non combo fixture id for staging area', () => {
-    const result = getMainNonComboFixtureId(
-      [
-        STAGING_AREA_RIGHT_SLOT_FIXTURE,
-        STAGING_AREA_SLOT_WITH_MAGNETIC_BLOCK_V1_FIXTURE,
-        STAGING_AREA_SLOT_WITH_WASTE_CHUTE_RIGHT_ADAPTER_COVERED_FIXTURE,
-        STAGING_AREA_SLOT_WITH_WASTE_CHUTE_RIGHT_ADAPTER_NO_COVER_FIXTURE,
-      ],
-      ['D4'],
-      'cutoutD3'
-    )
-    expect(result).toEqual(STAGING_AREA_RIGHT_SLOT_FIXTURE)
-  })
-
-  it('should get main non combo fixture id for flex stacker', () => {
-    const result = getMainNonComboFixtureId(
-      [
-        FLEX_STACKER_WITH_MAG_BLOCK_FIXTURE,
-        FLEX_STACKER_V1_FIXTURE,
-        FLEX_STACKER_WITH_WASTE_CHUTE_ADAPTER_NO_COVER_FIXTURE,
-        FLEX_STACKER_WITH_WASTE_CHUTE_ADAPTER_COVERED_FIXTURE,
-      ],
-      ['flexStackerModuleV1D4'],
-      'cutoutD3'
-    )
-    expect(result).toEqual(FLEX_STACKER_V1_FIXTURE)
-  })
-
-  it('should get main non combo fixture id for magnetic block', () => {
-    const result = getMainNonComboFixtureId(
-      [
-        FLEX_STACKER_WITH_MAG_BLOCK_FIXTURE,
-        MAGNETIC_BLOCK_V1_FIXTURE,
-        STAGING_AREA_SLOT_WITH_MAGNETIC_BLOCK_V1_FIXTURE,
-      ],
-      ['magneticBlockV1D3'],
-      'cutoutD3'
-    )
-    expect(result).toEqual(MAGNETIC_BLOCK_V1_FIXTURE)
   })
 })
