+++ shared-data/js/fixtures.ts
@@ -3,12 +3,9 @@
  * Not to be confused with fixtures in the sense of dummy test data.
  */
 import {
-  COMBO_FIXTURES,
   FAKE_STAGING_SLOT_WITH_MAG_BLOCK_FIXTURE,
-  FLEX_MODULE_AA_TYPE_BY_MODEL,
   FLEX_STACKER_FIXTURES,
   FLEX_STAGING_ADDRESSABLE_AREAS_WITH_FAKES,
-  FLEX_USB_MODULE_FIXTURES,
   THERMOCYCLER_MODULE_CUTOUTS,
   WASTE_CHUTE_CUTOUT,
   WASTE_CHUTE_FLEX_STACKER_FIXTURES,
@@ -77,12 +74,12 @@ import {
   getAAForModuleFixture,
   getAAsToFixtureIdFromDeckDefWithFakes,
   getAAWithFakesFromCutoutFixtureId,
-  getMainAAForAFixture,
 } from './helpers/deckConfiguration/getAddressableAreaFrom'
 import {
   getCutoutFixtureIdsForModuleModel,
   getCutoutFixturesForModuleModel,
   getFixtureIdByCutoutIdFromModuleSlotName,
+  getMainFixtureIdForAA,
   getReplacementFixtureForFakeFixture,
   replaceFixtureToFakeFixtureAndTransformCutoutFixturesToAA,
 } from './helpers/deckConfiguration/getFixtureFrom'
@@ -109,7 +106,6 @@ import type {
 import type { CutoutConfigAndCompatibility } from './helpers'
 import type {
   AddressableArea,
-  AreaType,
   CoordinateTuple,
   CutoutConfig,
   CutoutConfigMap,
@@ -733,89 +729,6 @@ export const isAddressableAreaStandardSlot = (
     ? STANDARD_FLEX_SLOTS
     : STANDARD_OT2_SLOTS
   ).includes(addressableAreaName)
-
-export const getMainUsbModuleFixtureIdForComboFixture = (
-  compatibleCutoutFixtureIds: CutoutFixtureId[]
-): CutoutFixtureId | null => {
-  return (
-    compatibleCutoutFixtureIds.find(cf =>
-      FLEX_USB_MODULE_FIXTURES.includes(cf)
-    ) ?? null
-  )
-}
-
-export const getMainFixtureIdForAA = (
-  compatibleCutoutFixtureIds: CutoutFixtureId[],
-  addressableAreaIds: AddressableAreaName[],
-  cutoutId: CutoutId
-): CutoutFixtureId | null => {
-  if (addressableAreaIds.length === 1) {
-    return getMainNonComboFixtureId(
-      compatibleCutoutFixtureIds,
-      addressableAreaIds,
-      cutoutId
-    )
-  }
-
-  const deckDef = getDeckDefFromRobotType('OT-3 Standard')
-  const cutoutFixtures = deckDef.cutoutFixtures.filter(cf =>
-    compatibleCutoutFixtureIds.includes(cf.id)
-  )
-  const cutoutFixturesWithAddressableAreas = cutoutFixtures.find(cf =>
-    Object.values(cf.providesAddressableAreas).some(providedAAs =>
-      addressableAreaIds.every(aa => providedAAs.includes(aa))
-    )
-  )
-  return cutoutFixturesWithAddressableAreas?.id ?? null
-}
-
-export const getMainNonComboFixtureId = (
-  compatibleCutoutFixtureIds: CutoutFixtureId[],
-  addressableAreaIds: AddressableAreaName[],
-  cutoutId: CutoutId
-): CutoutFixtureId | null => {
-  const deckDef = getDeckDefFromRobotType('OT-3 Standard')
-  const cutoutFixtures = deckDef.cutoutFixtures.filter(cf =>
-    compatibleCutoutFixtureIds.includes(cf.id)
-  )
-  const aaAreaType = getAAByAAId(addressableAreaIds[0], deckDef).areaType
-  if (
-    Object.values(FLEX_MODULE_AA_TYPE_BY_MODEL).includes(
-      aaAreaType as AreaType
-    ) &&
-    aaAreaType !== 'magneticBlock'
-  ) {
-    return (
-      getMainUsbModuleFixtureIdForComboFixture(compatibleCutoutFixtureIds) ??
-      null
-    )
-  }
-  const cutoutFixturesWithAddressableAreas = cutoutFixtures.filter(cf =>
-    Object.values(cf.providesAddressableAreas).some(providedAAs =>
-      addressableAreaIds.every(aa => providedAAs.includes(aa))
-    )
-  )
-
-  if (cutoutFixturesWithAddressableAreas.length === 0) {
-    return null
-  }
-
-  // Find the fixture with the least items in its providesAddressableAreas in order to find the simplest fixture
-  const fixtureWithLeastAAs = cutoutFixturesWithAddressableAreas.reduce(
-    (minFixture, currentFixture) => {
-      const minAAsCount = Object.entries(minFixture.providesAddressableAreas)
-        .filter(([key, value]) => key === cutoutId)
-        .map(([key, value]) => value)[0].length
-      const currentAAsCount = Object.entries(
-        currentFixture.providesAddressableAreas
-      )
-        .filter(([key, value]) => key === cutoutId)
-        .map(([key, value]) => value)[0].length
-      return currentAAsCount < minAAsCount ? currentFixture : minFixture
-    }
-  )
-  return fixtureWithLeastAAs.id
-}
 
 export const isModuleAllowedOnAA = (
   cutoutId: CutoutId,
